Consolidate application into single GUI script
